In the script.js think I didn' svae so had to redo
I made the game work, as got someone to play with it and
it worked. But when got it off my github, it was unfinished.
Now tried to use getComputerChoice() as second argument, comp
choice, so it's calling gCC() each time I call rPS(), and it
woks.computerChoice variable is also set to gCC(), so will
try and put it as an argument in calling rPS(), and just will place
gCC() in first line of rPS()to see if that works as well, meaning
if it will reset random number each calling, as it is now.